Reshuffle test order
