+++ jol-samples/src/main/java/org/openjdk/jol/samples/JOLSample_20_Allocation.java
@@ -30,7 +30,8 @@
  */
 package org.openjdk.jol.samples;
 
-import org.openjdk.jol.info.ClassLayout;
+
+
 import org.openjdk.jol.vm.VM;
 
 import java.io.PrintWriter;
@@ -40,57 +41,44 @@ import static java.lang.System.out;
 /**
  * @author Aleksey Shipilev
  */
-public class JOLSample_19_Promotion {
+public class JOLSample_20_Allocation {
 
     /*
-     * The example of object promotion.
+     * The example of allocation addresses.
      *
-     * Once the object survives the garbage collections, it is getting
-     * promoted to another generation. In this example, we can track
-     * the addresses of the objects, as it changes over time.
+     * This example shows the addresses of newly allocated objects
+     * grow linearly in HotSpot. This is because the allocation in
+     * parallel collectors is linear. We can also see it rewinds back
+     * to the same offsets -- that's the start of some GC generation.
      *
-     * VM also needs to record the "age" (that is, the number of GC
-     * cycles the object had survived) of the object somewhere, and
-     * it is stored in mark word as well. See how particular mark word
-     * bits change with each promotion.
+     * For Parallel-like GCs, while GC adjusts for the allocation rate.
+     * For G1-like GCs, the allocation address changes by region size,
+     * as collector switches to another region for allocation.
      *
      * Run with test with smaller heap (about 1 GB) for best results.
      */
 
-    static volatile Object sink;
-
     public static void main(String[] args) {
         out.println(VM.current().details());
 
-        PrintWriter pw = new PrintWriter(System.out, true);
+        PrintWriter pw = new PrintWriter(out, true);
 
-        Object o = new Object();
+        long last = VM.current().addressOf(new Object());
+        for (int l = 0; l < 1000 * 1000 * 1000; l++) {
+            long current = VM.current().addressOf(new Object());
 
-        ClassLayout layout = ClassLayout.parseInstance(o);
-
-        long lastAddr = VM.current().addressOf(o);
-        pw.printf("*** Fresh object is at %x%n", lastAddr);
-        out.println(layout.toPrintable());
-
-        int moves = 0;
-        for (int i = 0; i < 100000; i++) {
-            long cur = VM.current().addressOf(o);
-            if (cur != lastAddr) {
-                moves++;
-                pw.printf("*** Move %2d, object is at %x%n", moves, cur);
-                out.println(layout.toPrintable());
-                lastAddr = cur;
+            long distance = Math.abs(current - last);
+            if (distance > 4096) {
+                pw.printf("Jumping from %x to %x (distance = %d bytes, %dK, %dM)%n",
+                        last,
+                        current,
+                        distance,
+                        distance / 1024,
+                        distance / 1024 / 1024);
             }
 
-            // make garbage
-            for (int c = 0; c < 10000; c++) {
-                sink = new Object();
-            }
+            last = current;
         }
-
-        long finalAddr = VM.current().addressOf(o);
-        pw.printf("*** Final object is at %x%n", finalAddr);
-        out.println(layout.toPrintable());
 
         pw.close();
     }
