+++ jol-samples/src/main/java/org/openjdk/jol/samples/JOLSample_25_Compaction.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, Red Hat, Inc.
+ * Copyright (c) 2014, Oracle America, Inc.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -30,73 +30,55 @@
  */
 package org.openjdk.jol.samples;
 
-import org.openjdk.jol.info.ClassLayout;
+import org.openjdk.jol.info.GraphLayout;
 import org.openjdk.jol.vm.VM;
 
-import java.util.concurrent.TimeUnit;
+import java.util.ArrayList;
+import java.util.List;
 
 import static java.lang.System.out;
 
 /**
  * @author Aleksey Shipilev
  */
-public class JOLSample_26_IHC_BL_Conflict {
+public class JOLSample_25_Compaction {
 
     /*
-     * This is the example of biased locking conflicting with identity hash
-     * code. Identity hash code takes precedence.
+     * This is the example how VM compacts the objects.
      *
-     * In order to demonstrate this, we first need to sleep for >5 seconds
-     * to pass the grace period of biased locking. Then, we do the same
-     * trick as the example before. You may notice that the mark word
-     * had not changed after the first lock was released, retaining the bias.
+     * This example generates PNG images in your current directory.
      *
-     * The identity hash code computation overwrites the biased locking information,
-     * and subsequent locks only displace it temporarily. After the second lock
-     * is released, identity hash code data gets back. No biased locking is
-     * possible for that object anymore.
+     * You can see the freshly allocated and populated list has quite
+     * the sparse layout. It happens because many temporary objects are
+     * allocated while populating the list. The subsequent GCs compact
+     * the list into the one or few dense blocks.
      *
-     * On JDK 15+, this test should enable -XX:+UseBiasedLocking.
+     * Run this test with -Xmx1g -XX:+UseParallelGC -XX:ParallelGCThreads=1
+     * for best results.
      */
+
+    public static volatile Object sink;
 
     public static void main(String[] args) throws Exception {
         out.println(VM.current().details());
 
-        TimeUnit.SECONDS.sleep(6);
-
-        final A a = new A();
-
-        ClassLayout layout = ClassLayout.parseInstance(a);
-
-        out.println("**** Fresh object");
-        out.println(layout.toPrintable());
-
-        synchronized (a) {
-            out.println("**** With the lock");
-            out.println(layout.toPrintable());
+        // allocate some objects to beef up generations
+        for (int c = 0; c < 1000000; c++) {
+            sink = new Object();
         }
 
-        out.println("**** After the lock");
-        out.println(layout.toPrintable());
+        System.gc();
 
-        int hashCode = a.hashCode();
-        out.println("hashCode: " + Integer.toHexString(hashCode));
-        out.println();
-
-        out.println("**** After the hashcode");
-        out.println(layout.toPrintable());
-
-        synchronized (a) {
-            out.println("**** With the second lock");
-            out.println(layout.toPrintable());
+        List<String> list = new ArrayList<>();
+        for (int c = 0; c < 1000; c++) {
+            list.add("Key" + c);
         }
 
-        out.println("**** After the second lock");
-        out.println(layout.toPrintable());
+        for (int c = 1; c <= 10; c++) {
+            GraphLayout.parseInstance(list).toImage("list-" + c + ".png");
+            System.gc();
+        }
     }
 
-    public static class A {
-        // no fields
-    }
 
 }
